+++ src/util/xpub.rs
@@ -5,12 +5,12 @@ use std::str::FromStr;
 pub use serde::de;
 
 use bitcoin::util::bip32::{ChildNumber, DerivationPath, ExtendedPubKey, Fingerprint};
-use bitcoin::{util::base58, Address, Network};
-use miniscript::descriptor::{Descriptor, DescriptorPublicKey, DescriptorXPub};
+use bitcoin::{util::base58, Network};
+use miniscript::descriptor::{Descriptor, DescriptorPublicKey, DescriptorXKey};
 
 use crate::types::ScriptType;
 use crate::util::descriptor::ExtendedDescriptor;
-use crate::util::{BoolThen, EC};
+use crate::util::BoolThen;
 
 pub fn xpub_matches_network(xpub: &ExtendedPubKey, network: Network) -> bool {
     // testnet and regtest share the same bip32 version bytes
@@ -40,9 +40,9 @@ impl XyzPubKey {
             )
         });
 
-        let desc_key = DescriptorPublicKey::XPub(DescriptorXPub {
+        let desc_key = DescriptorPublicKey::XPub(DescriptorXKey {
             origin: bip32_origin,
-            xpub: self.xpub,
+            xkey: self.xpub,
             derivation_path,
             is_wildcard: true,
         });
@@ -51,40 +51,6 @@ impl XyzPubKey {
             ScriptType::P2pkh => Descriptor::Pkh(desc_key),
             ScriptType::P2wpkh => Descriptor::Wpkh(desc_key),
             ScriptType::P2shP2wpkh => Descriptor::ShWpkh(desc_key),
-        }
-    }
-
-    /// Convert simple p2*pkh ranged descriptors to their XyzPubKey representation
-    pub fn try_from_desc(desc: &ExtendedDescriptor) -> Option<Self> {
-        let (script_type, desc_xpub) = match desc {
-            Descriptor::Pkh(DescriptorPublicKey::XPub(xpub)) => (ScriptType::P2pkh, xpub),
-            Descriptor::Wpkh(DescriptorPublicKey::XPub(xpub)) => (ScriptType::P2wpkh, xpub),
-            Descriptor::ShWpkh(DescriptorPublicKey::XPub(xpub)) => (ScriptType::P2shP2wpkh, xpub),
-            _ => return None,
-        };
-
-        if !desc_xpub.is_wildcard {
-            return None;
-        }
-
-        Some(XyzPubKey {
-            script_type,
-            xpub: desc_xpub
-                .xpub
-                .derive_pub(&*EC, &desc_xpub.derivation_path)
-                .unwrap(),
-        })
-    }
-
-    /// Get the address of the key at the specified derivation index
-    /// Panics if given a hardened child number
-    pub fn derive_address(&self, index: u32, network: Network) -> Address {
-        let key = self.xpub.ckd_pub(&*EC, index.into()).unwrap();
-        match self.script_type {
-            ScriptType::P2pkh => Address::p2pkh(&key.public_key, network),
-            ScriptType::P2wpkh => Address::p2wpkh(&key.public_key, network).unwrap(),
-            ScriptType::P2shP2wpkh => Address::p2shwpkh(&key.public_key, network).unwrap(),
-            // the two unwraps above can only fail if the public key is non-compressed, which it cannot be.
         }
     }
 }
@@ -191,10 +157,8 @@ fn get_xpub_p2pkh_version(network: Network) -> [u8; 4] {
 mod tests {
     use super::*;
 
-    // Test xyzpub -> descriptor -> xyzpub roundtrip
     #[test]
-    fn test_xpub_to_desc_conversion() {
-        let net = Network::Bitcoin;
+    fn test_xyzpub_to_desc() {
         let test_cases = [
             // Standard BIP32 xpub, uses p2pkh
             ("xpub661MyMwAqRbcFLqTBCNzuoj4FYE1xRxmCjrSWC6LUjKHo46Du4NacKgxdrJPWhzLjkPsXqnjAUwn1raMSWfxWZKysPoBNQMZMs8b5JM8egC",
@@ -211,56 +175,8 @@ mod tests {
         for (xyz_str, expected_desc) in &test_cases {
             let xyzpub = xyz_str.parse::<XyzPubKey>().unwrap();
             let desc = xyzpub.as_descriptor([][..].into());
-            let xyzpub_rt = XyzPubKey::try_from_desc(&desc).unwrap();
 
             assert_eq!(desc.to_string(), *expected_desc);
-            assert_eq!(xyzpub_rt.xpub, xyzpub.xpub);
-            assert_eq!(xyzpub_rt.script_type, xyzpub.script_type);
-
-            let address = xyzpub.derive_address(9, net);
-            assert_eq!(desc.derive(9.into()).address(net).unwrap(), address);
-            assert_eq!(xyzpub_rt.derive_address(9, net), address);
-        }
-    }
-
-    // Test descriptor -> xyzpub -> descriptor roundtrip
-    #[test]
-    fn test_desc_to_xpub_conversion() {
-        let net = Network::Bitcoin;
-        // Simple ranged p2*pkh descriptors that can be represented as as optimized XyzPubKey
-        let test_cases =[
-            // p2pkh
-            ("pkh(xpub661MyMwAqRbcFLqTBCNzuoj4FYE1xRxmCjrSWC6LUjKHo46Du4NacKgxdrJPWhzLjkPsXqnjAUwn1raMSWfxWZKysPoBNQMZMs8b5JM8egC/*)",
-             "xpub661MyMwAqRbcFLqTBCNzuoj4FYE1xRxmCjrSWC6LUjKHo46Du4NacKgxdrJPWhzLjkPsXqnjAUwn1raMSWfxWZKysPoBNQMZMs8b5JM8egC",
-             ScriptType::P2pkh),
-
-            // p2wpkh with xpub child derivation
-            ("wpkh(xpub661MyMwAqRbcFLqTBCNzuoj4FYE1xRxmCjrSWC6LUjKHo46Du4NacKgxdrJPWhzLjkPsXqnjAUwn1raMSWfxWZKysPoBNQMZMs8b5JM8egC/0/*)",
-             "xpub68VHDuZRhKBTDwzEiVPAL8gfPvkLQiUYsZ4W7PAT6LxPYchGuSXh7NQBL418maAsf89gZsDTntQVzPC37qmxd3qKvJMbAGCSV5eBjUwiPZk",
-             ScriptType::P2wpkh),
-        ];
-        for (desc_str, expected_xpub, expected_type) in &test_cases {
-            let desc = desc_str.parse::<ExtendedDescriptor>().unwrap();
-            let xyzpub = XyzPubKey::try_from_desc(&desc).unwrap();
-            let desc_rt = xyzpub.as_descriptor([][..].into());
-
-            assert_eq!(xyzpub.xpub.to_string(), *expected_xpub);
-            assert_eq!(xyzpub.script_type, *expected_type);
-
-            let address = desc.derive(9.into()).address(net).unwrap();
-            assert_eq!(xyzpub.derive_address(9, net), address);
-            assert_eq!(desc_rt.derive(9.into()).address(net).unwrap(), address);
-        }
-
-        // Descriptors without an XyzPubKey representation
-        let unoptimizable_descs = [
-          "wsh(multi(1,tpubD6NzVbkrYhZ4XmWGpWP6vdR1uS1NVvgUgM3wFUzCywE8nupMQpmvBGBYzjcZfHX46xSCpBxmFSswJzE98vsL48hW5HsampQhRBnKUHin36y/*))",
-          // non-ranged, no child derivation to optimize
-          "pkh(tpubD6NzVbkrYhZ4XmWGpWP6vdR1uS1NVvgUgM3wFUzCywE8nupMQpmvBGBYzjcZfHX46xSCpBxmFSswJzE98vsL48hW5HsampQhRBnKUHin36y)",
-          "pkh(021ebb0d349ccd72d3648c944c84e38345cf8d200dcf216cb624a0b869bbf974f0)",
-        ];
-        for desc_str in &unoptimizable_descs {
-            assert!(XyzPubKey::try_from_desc(&desc_str.parse().unwrap()).is_none());
         }
     }
 }